DXIndicator moved to adx-package
